+++ src/components/ui/main.js
@@ -2,3 +2,4 @@ export * from './Loading';
 export * from './ProgressBar';
 export * from './SidebarToggle/index';
 export * from './ValidateSaved';
+export * from './SimpleProgressBar';